+++ src/main/javascript/src/lib/constants.ts
